refactor

+++ drawwing/src/draww/axes.hpp
@@ -4,7 +4,7 @@
 #include "linalg.hpp"
 
 struct Axis {
-	float center;  // absolute
+	double center;  // absolute
 	int scale;  // pixels per unit
 };
 
@@ -17,20 +17,20 @@ public:
 	CoordinateSystem(Axis x_ax, Axis y_ax, SDL_FRect dims)
 		: x_axis(x_ax), y_axis(y_ax), dim(dims) {}
 
-	float x_screen_to_space(float x_screen) const {
-		return (x_screen - y_axis.center) / (float)x_axis.scale;
+	double x_screen_to_space(double x_screen) const {
+		return (x_screen - y_axis.center) / (double)x_axis.scale;
 	}
 
-	float y_screen_to_space(float y_screen) const {
-		return (x_axis.center - y_screen) / (float)y_axis.scale;
+	double y_screen_to_space(double y_screen) const {
+		return (x_axis.center - y_screen) / (double)y_axis.scale;
 	}
 
-	float x_space_to_screen(float x_space) const {
-		return y_axis.center + x_space * (float)x_axis.scale;
+	double x_space_to_screen(double x_space) const {
+		return y_axis.center + x_space * (double)x_axis.scale;
 	}
 
-	float y_space_to_screen(float y_space) const {
-		return x_axis.center - y_space * (float)y_axis.scale;
+	double y_space_to_screen(double y_space) const {
+		return x_axis.center - y_space * (double)y_axis.scale;
 	}
 
 	void vector2_space_to_screen(Vector2 *vec) const {
